update: reorganized components default and named exports
