+++ src/components/Button/Button.tsx
@@ -5,12 +5,10 @@ interface ButtonProps {
   action: () => void;
 }
 
-const Button: React.FC<ButtonProps> = ({ text, action }) => {
+export const Button: React.FC<ButtonProps> = ({ text, action }) => {
   return (
     <button onClick={action} className={styles.btn}>
       {text}
     </button>
   );
 };
-
-export default Button;
