+++ src/components/Loader/Loader.tsx
@@ -1,7 +1,5 @@
 import styles from "./Loader.module.scss";
 
-const Loader = () => {
+export const Loader = () => {
   return <div className={styles.loader} role="status"></div>;
 };
-
-export default Loader;
