+++ src/components/Logo/Logo.tsx
@@ -1,8 +1,6 @@
 import logo from "../../assets/logo.svg";
 import styles from "./Logo.module.scss";
 
-const Logo: React.FC = () => (
+export const Logo: React.FC = () => (
   <img src={logo} alt="Nature Lovers" loading="lazy" className={styles.logo} />
 );
-
-export default Logo;
