+++ src/components/__tests__/ResponsiveImage.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import { vi } from "vitest";
 import { RESPONSIVE_SIZES } from "../../constants/responsiveSizes";
 import { ImageUrls, Photo } from "../../types/imageTypes";
-import ResponsiveImage from "../ResponsiveImage/ResponsiveImage";
+import { ResponsiveImage } from "../index";
 import { getImageUrls } from "../../utils/getImageUrls";
 
 vi.mock("../../utils/getImageUrls", () => ({
