+++ src/components/index.ts
@@ -1,0 +1,9 @@
+export * from "./Button/Button";
+export * from "./Error/Error";
+export * from "./ImageCard/ImageCard";
+export * from "./ImageDetails/ImageDetails";
+export * from "./Loader/Loader";
+export * from "./Logo/Logo";
+export * from "./Main/Main";
+export * from "./ResponsiveImage/ResponsiveImage";
+export * from "./Tabs/Tabs";
