+++ src/pages/Home/sections/Hero/Hero.tsx
@@ -1,5 +1,4 @@
-import Logo from "../../../../components/Logo/Logo";
-import Tabs, { TabsType } from "../../../../components/Tabs/Tabs";
+import { Logo, Tabs, TabsType } from "../../../../components/index";
 import styles from "./Hero.module.scss";
 
 interface HeroProps {
@@ -7,11 +6,9 @@ interface HeroProps {
   handleActiveTab: (tab: TabsType) => void;
 }
 
-const Hero: React.FC<HeroProps> = ({ activeTab, handleActiveTab }) => (
+export const Hero: React.FC<HeroProps> = ({ activeTab, handleActiveTab }) => (
   <div className={styles.hero}>
     <Logo />
     <Tabs activeTab={activeTab} handleActiveTab={handleActiveTab} />
   </div>
 );
-
-export default Hero;
