+++ src/pages/Home/sections/index.ts
@@ -1,0 +1,4 @@
+export * from "./FavouriteList/FavouriteList";
+export * from "./Hero/Hero";
+export * from "./ImagesFilter/ImagesFilter";
+export * from "./ImagesList/ImagesList";
